+++ packages/common/tokens/appSwitcher.ts
@@ -1,0 +1,3 @@
+export const appSwitcherWidth = 359;
+export const denseAppSwitcherWidth = 324;
+export const appSwitcherMaxSize = 600;